requirements_frozen.txt
absl-py==0.14.0
asciitree==0.3.3
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
decorator==5.1.0
docker-pycreds==0.4.0
fasteners==0.16.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
ipython==7.28.0
jedi==0.18.0
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.37.0
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
msgpack==1.0.2
numba==0.54.0
numcodecs==0.9.1
numpy==1.19.2
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prettytable==2.2.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
pudb==2021.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pynndescent==0.5.4
pyparsing==2.4.7
python-box==5.4.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
sentry-sdk==1.4.1
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.26.7
urwid==2.1.2
wandb==0.12.2
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yaspin==2.1.0
zarr==2.10.0